requirements.txt

Django
djangorestframework
pytz
drf-yasg